Add Garnet.shared project and move LightEpoch into it
Introduce libs/shared/Garnet.shared.csproj as a low-level project shared
by Tsavorite and the Garnet libraries, and move LightEpoch (plus its
EntryTable/TestHooks partials and IEpochAccessor) out of Tsavorite.core
into it under the Garnet.shared namespace.

Murmur3 moves to Garnet.shared.HashUtility so LightEpoch has no
dependency on Tsavorite.core; Tsavorite.core.Utility.Murmur3 delegates
to it to preserve the public API.

The LightEpoch tests move from Tsavorite.test.epoch to a new
test/standalone/Garnet.shared.test project, which is added to the CI and
nightly standalone test matrices so the epoch tests actually run in CI.

Copilot-Session: ba1caae5-c8f5-43b4-be06-a73874807ea7
