+++ libs/server/Garnet.server.csproj
@@ -34,6 +34,7 @@
 
 	<!-- Include shared global usings file for server projects -->
 	<ItemGroup>
+		<Using Include="Garnet.shared" />
 		<Compile Include="$(MSBuildThisFileDirectory)../GlobalUsings.cs" Link="GlobalUsings.cs" />
 	</ItemGroup>
 </Project>
