+++ libs/shared/Epochs/LightEpoch.cs
@@ -9,7 +9,7 @@ using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 using System.Threading;
 
-namespace Tsavorite.core
+namespace Garnet.shared
 {
     /// <summary>
     /// Epoch protection
@@ -665,7 +665,7 @@ namespace Tsavorite.core
             if (Metadata.threadId == 0) // run once per thread for performance
             {
                 Metadata.threadId = Environment.CurrentManagedThreadId;
-                var code = (uint)Utility.Murmur3(Metadata.threadId);
+                var code = (uint)HashUtility.Murmur3(Metadata.threadId);
                 Metadata.startOffset1 = (ushort)(1 + (code % kTableSize));
                 Metadata.startOffset2 = (ushort)(1 + ((code >> 16) % kTableSize));
             }
