+++ test/standalone/Garnet.shared.test/Garnet.shared.test.csproj
@@ -2,7 +2,7 @@
 
   <PropertyGroup>
     <SignAssembly>true</SignAssembly>
-    <AssemblyOriginatorKeyFile>../../../../../../Garnet.snk</AssemblyOriginatorKeyFile>
+    <AssemblyOriginatorKeyFile>../../../Garnet.snk</AssemblyOriginatorKeyFile>
     <DelaySign>false</DelaySign>
     <IsPackable>false</IsPackable>
     <NoWarn>1701;1702;1591;IDE0130;IDE0065;IDE0007;IDE0048</NoWarn>
@@ -18,8 +18,8 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\src\core\Tsavorite.core.csproj" />
-    <ProjectReference Include="..\..\..\..\..\..\test\standalone\Garnet.test\Garnet.test.csproj" />
+    <ProjectReference Include="..\..\..\libs\shared\Garnet.shared.csproj" />
+    <ProjectReference Include="..\Garnet.test\Garnet.test.csproj" />
   </ItemGroup>
 
 </Project>
